simple_service/app/workers/requirements.txt
pika==1.2.0
pandas==2.2.2
scikit-learn==1.4.1.post1
joblib==1.4.2
asyncpg==0.28.0
psycopg==3.1.10
psycopg-binary==3.1.10
alembic==1.12.0
pydantic==2.3.0
pydantic-settings==2.0.3
sqlmodel==0.0.19
sqlalchemy==2.0.31
starlette==0.37.2
tenacity==8.2.2
requests==2.32.3
passlib==1.7.4
python-jose==3.3.0
wandb==0.15.5
minio==7.2.15
Pillow==9.5.0
matplotlib==3.8.0
